Set latest flags and attention spec (07f52fe)
also pipes through a classifier free option properly for turbo parity with diffusers.

+++ models/turbine_models/custom_models/sdxl_inference/sdxl_cmd_opts.py
@@ -154,7 +154,7 @@ p.add_argument(
 p.add_argument(
     "--precision",
     type=str,
-    default="fp32",
+    default="fp16",
     help="Precision of Stable Diffusion weights and graph.",
 )
 p.add_argument(
